Ordering fixes
