+++ Quamotion.TurboJpegWrapper/TJDecompressor.cs
@@ -14,9 +14,9 @@ namespace TurboJpegWrapper
     /// </summary>
     public class TJDecompressor : IDisposable
     {
+        private readonly object @lock = new object();
         private IntPtr decompressorHandle = IntPtr.Zero;
         private bool isDisposed;
-        private readonly object @lock = new object();
 
         /// <summary>
         /// Initializes a new instance of the <see cref="TJDecompressor"/> class.
@@ -32,6 +32,14 @@ namespace TurboJpegWrapper
             {
                 TJUtils.GetErrorAndThrow();
             }
+        }
+
+        /// <summary>
+        /// Finalizes an instance of the <see cref="TJDecompressor"/> class.
+        /// </summary>
+        ~TJDecompressor()
+        {
+            this.Dispose(false);
         }
 
         /// <summary>
@@ -260,16 +268,6 @@ namespace TurboJpegWrapper
             return stride * height;
         }
 
-        private ColorPalette FixPaletteToGrayscale(ColorPalette palette)
-        {
-            for (var index = 0; index < palette.Entries.Length; ++index)
-            {
-                palette.Entries[index] = Color.FromArgb(index, index, index);
-            }
-
-            return palette;
-        }
-
         /// <summary>
         /// Releases resources.
         /// </summary>
@@ -293,6 +291,16 @@ namespace TurboJpegWrapper
             }
         }
 
+        private ColorPalette FixPaletteToGrayscale(ColorPalette palette)
+        {
+            for (var index = 0; index < palette.Entries.Length; ++index)
+            {
+                palette.Entries[index] = Color.FromArgb(index, index, index);
+            }
+
+            return palette;
+        }
+
         private void Dispose(bool callFromUserCode)
         {
             if (callFromUserCode)
@@ -310,14 +318,6 @@ namespace TurboJpegWrapper
                 // (i.e. make calling Dispose twice a safe thing to do).
                 this.decompressorHandle = IntPtr.Zero;
             }
-        }
-
-        /// <summary>
-        /// Finalizes an instance of the <see cref="TJDecompressor"/> class.
-        /// </summary>
-        ~TJDecompressor()
-        {
-            this.Dispose(false);
         }
     }
 }
